AI movement
Slight rework on collision system.
Slight rework on how to check for valid gameObjects
Slight rework to the movement system, changing movement to hold instead of press and keep moving.
Added point gaining when digging.
Player dies when touched by pooka,
resets to start pos
Have to add state to scenes to pause when this happens